Abstract Identity interfaces and classes to their own file
